requirements.txt
adafruit-circuitpython-ads1x15==2.2.22
gpiozero==1.6.2
RPi.GPIO==0.7.1
sensirion-i2c-driver==1.0.0
sensirion-i2c-scd==0.1.2

nicegui==1.3.13
psutil==5.9.5
requests==2.31.0
